+++ doc/mkdocs/theme/overrides/main.html
@@ -1,0 +1,12 @@
+{#-
+    This file was automatically generated - do not edit
+  -#}
+  {% extends "base.html" %}
+  {% block extrahead %}
+    <link rel="stylesheet" href="{{ 'overrides/assets/stylesheets/main.bf3dc0a9.min.css' | url }}">
+  {% endblock %}
+  {% block scripts %}
+    {{ super() }}
+    <script src="{{ 'overrides/assets/javascripts/bundle.525231ca.min.js' | url }}"></script>
+  {% endblock %}
+
